chore: rename SwiftLoggable to DebugLoggable
