+++ Airwallex/Payment/Sources/CollectionViewManager.swift
@@ -148,7 +148,7 @@ class CollectionViewManager<SectionType: Hashable & Sendable, ItemType: Hashable
     }
 }
 
-class CollectionViewContext<Section: Hashable & Sendable, Item: Hashable & Sendable>: SwiftLoggable {
+class CollectionViewContext<Section: Hashable & Sendable, Item: Hashable & Sendable>: DebugLoggable {
     private(set) weak var viewController: UIViewController?
     private weak var collectionView: UICollectionView!
     private weak var layout: UICollectionViewCompositionalLayout!
